Check if a PR already exists before pushing the branch
Fixes https://github.com/foolip/safari-technology-preview-updater/issues/4.
